requirements.txt
alembic==1.8.1
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
dnspython==2.2.1
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.85.1
greenlet==2.0.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
Mako==1.2.3
MarkupSafe==2.1.1
orjson==3.8.1
packaging==21.3
passlib==1.7.4
Pillow==9.4.0
pipwin==0.5.2
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
pytest==7.2.0
pytest-asyncio==0.20.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlmodel==0.0.8
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.18.0
websockets==10.4
